% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIGOR.R
\docType{package}
\name{LIGOR}
\alias{LIGOR}
\alias{LIGOR-package}
\title{LIGOR package}
\description{
\code{LIGOR} is a package used
to intepret the LIGO data in R. It provides the following functions \code{\link{loaddata}},
\code{\link{whiten}} and \code{\link{loadtemplate}}.
}
\seealso{
\code{\link{loaddata}}

\code{\link{whiten}}

\code{\link{loadtemplate}}
}
