% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaddata.R
\name{loaddata}
\alias{loaddata}
\title{Read the event properties from a local xml file}
\usage{
loaddata(filename, ifo = None)
}
\arguments{
\item{filename}{the name of the xml file to be read}

\item{ifo}{to specify which detector we are looking at, namely \code{H1} or \code{L1}}
}
\description{
Read the event properties from a local xml file
}
\keyword{data}
