% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadtemplate.R
\name{loadtemplate}
\alias{loadtemplate}
\title{Read template properties from a local xml file}
\usage{
loadtemplate(filename)
}
\arguments{
\item{filename}{the name of the xml file to be read}
}
\description{
Read template properties from a local xml file
}
\keyword{data}
