% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whiten.R
\name{whiten}
\alias{whiten}
\title{Whiten the data}
\usage{
whiten(strain, interp_psd, dt)
}
\arguments{
\item{strain}{the strain of a given detector, namely \code{H1} or \code{L1}}

\item{interp_psd}{the corresponding template}

\item{dt}{the frequency step size}
}
\description{
Divide the data by the noise amplitude spectrum in the Fourier domain
}
\keyword{ts}
